there is a ad placeholder at the bottom can we put adsence ad there. ?
i got this code from adsence

Copy and paste the ad unit code in between the <body></body> tags of your pages
Place this code where you want an ad to appear. Do this for each individual ad unit, on every page.

<script async src="https://pagead2.googlesyndication.com/pagead/js/adsbygoogle.js?client=ca-pub-7546609873197379"
     crossorigin="anonymous"></script>
<!-- footer-ad -->
<ins class="adsbygoogle"
     style="display:block"
     data-ad-client="ca-pub-7546609873197379"
     data-ad-slot="5097284517"
     data-ad-format="auto"
     data-full-width-responsive="true"></ins>
<script>
     (adsbygoogle = window.adsbygoogle || []).push({});
</script>
It usually takes a few minutes for ads to appear on the page but occasionally it can take up to an hour. See our code implementation guide⁠‎ for more details.